requirements.txt
matplotlib>=3.3.0
numpy>=1.18.0
paho-mqtt>=1.6.1
pandas>=1.0
python-dotenv
scikit-learn>=0.22
scipy>=1.4
tk
torch
